test: CategoryService 불필요한 테스트 제거 후 CategoryFacadeService 테스트 추가
- CategoryService에서는 단순히 전체 Category 엔티티 조회하는 로직만 존재하므로 테스트 제거
- CategoryFacadeService에서는 전체 Category 엔티티 조회 후 dto로 변환하는 로직 포함 → 전체 카테고리 조회 테스트 추가

[SW-30]
