+++ src/test/java/com/wanted/safewallet/domain/category/business/service/CategoryServiceTest.java
@@ -4,7 +4,6 @@ import static com.wanted.safewallet.domain.category.persistence.entity.CategoryT
 import static com.wanted.safewallet.domain.category.persistence.entity.CategoryType.TRAFFIC;
 import static com.wanted.safewallet.global.exception.ErrorCode.NOT_FOUND_CATEGORY;
 import static com.wanted.safewallet.utils.Fixtures.aCategory;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
@@ -31,26 +30,6 @@ class CategoryServiceTest {
 
     @Mock
     CategoryRepository categoryRepository;
-
-    @DisplayName("전체 카테고리 조회 테스트 : 성공")
-    @Test
-    void getCategoryList() {
-        //given
-        List<Category> categoryList = List.of(aCategory().id(1L).type(FOOD).build(),
-            aCategory().id(2L).type(TRAFFIC).build());
-        given(categoryRepository.findAll()).willReturn(categoryList);
-
-        //when
-        List<Category> findedCategoryList = categoryService.getCategoryList();
-
-        //then
-        then(categoryRepository).should(times(1)).findAll();
-        assertThat(findedCategoryList).hasSize(2);
-        assertThat(findedCategoryList)
-            .extracting("id").contains(1L, 2L);
-        assertThat(findedCategoryList)
-            .extracting("type").contains(FOOD, TRAFFIC);
-    }
 
     @DisplayName("카테고리 유효성 검증 테스트 : 실패 - 존재하지 않는 카테고리")
     @Test
